APPLIED THIS ALL TO: EaganTest2
Added DeathZone prefab
Added DeathCondition script to kill the player if they enter a deathzone
modified the main loop song to .60f instead of 1.0f as it was too loud for playing the other instruments. (sounds great by the way)
modified sound script to ensure the first loop song won't play up to 1.0f (as we are using it twice TODO: remove the sound object that is playing the first song)
Modified level 001 to allow for BAD paths that can contain the prefab deathzone within the scene in those locations.
modified playerController to allow for some public members to allow for the deathZones to work properly.
Modified the rumbleRock sound script to better play once the rocks crumble. TODO: add a better sound over the top of the old crumble sound to not sound like rolling dice)